feat: add new git providers
