+++ apps/dokploy/server/api/routers/admin.ts
@@ -86,96 +86,9 @@ export const adminRouter = createTRPCRouter({
 			}
 		}),
 
-	cleanGithubApp: adminProcedure.mutation(async ({ ctx }) => {
-		try {
-			return await updateAdmin(ctx.user.authId, {
-				githubAppName: "",
-				githubClientId: "",
-				githubClientSecret: "",
-				githubInstallationId: "",
-			});
-		} catch (error) {
-			throw new TRPCError({
-				code: "BAD_REQUEST",
-				message: "Error to delete this github app",
-				cause: error,
-			});
-		}
-	}),
+	// haveGithubConfigured: protectedProcedure.query(async () => {
+	// 	const adminResponse = await findAdmin();
 
-	getRepositories: protectedProcedure
-		.input(
-			z.object({
-				githubProviderId: z.string().optional(),
-			}),
-		)
-		.query(async ({ input }) => {
-			if (!input.githubProviderId) {
-				return [];
-			}
-
-			const githubProvider = await getGithubProvider(input.githubProviderId);
-
-			const octokit = new Octokit({
-				authStrategy: createAppAuth,
-				auth: {
-					appId: githubProvider.githubAppId,
-					privateKey: githubProvider.githubPrivateKey,
-					installationId: githubProvider.githubInstallationId,
-				},
-			});
-
-			const repositories = (await octokit.paginate(
-				octokit.rest.apps.listReposAccessibleToInstallation,
-			)) as unknown as Awaited<
-				ReturnType<typeof octokit.rest.apps.listReposAccessibleToInstallation>
-			>["data"]["repositories"];
-
-			return repositories;
-		}),
-	getBranches: protectedProcedure
-		.input(apiGetBranches)
-		.query(async ({ input }) => {
-			// const admin = await findAdmin();
-
-			// const completeRequirements = haveGithubRequirements(admin);
-
-			// if (!completeRequirements) {
-			// 	throw new TRPCError({
-			// 		code: "BAD_REQUEST",
-			// 		message: "Admin need to setup correctly github account",
-			// 	});
-			// }
-
-			if (!input.githubProviderId) {
-				return [];
-			}
-			const githubProvider = await getGithubProvider(input.githubProviderId);
-
-			const octokit = new Octokit({
-				authStrategy: createAppAuth,
-				auth: {
-					appId: githubProvider.githubAppId,
-					privateKey: githubProvider.githubPrivateKey,
-					installationId: githubProvider.githubInstallationId,
-				},
-			});
-
-			const branches = (await octokit.paginate(
-				octokit.rest.repos.listBranches,
-				{
-					owner: input.owner,
-					repo: input.repo,
-				},
-			)) as unknown as Awaited<
-				ReturnType<typeof octokit.rest.repos.listBranches>
-			>["data"];
-
-			return branches;
-		}),
-	haveGithubConfigured: protectedProcedure.query(async () => {
-		const adminResponse = await findAdmin();
-
-		return haveGithubRequirements(adminResponse);
-	}),
+	// 	return haveGithubRequirements(adminResponse);
+	// }),
 });
