+++ apps/dokploy/server/db/schema/admin.ts
@@ -1,5 +1,5 @@
 import { relations } from "drizzle-orm";
-import { boolean, integer, pgTable, text } from "drizzle-orm/pg-core";
+import { boolean, pgTable, text } from "drizzle-orm/pg-core";
 import { createInsertSchema } from "drizzle-zod";
 import { nanoid } from "nanoid";
 import { z } from "zod";
@@ -13,17 +13,9 @@ export const admins = pgTable("admin", {
 		.notNull()
 		.primaryKey()
 		.$defaultFn(() => nanoid()),
-
-	githubAppId: integer("githubAppId"),
-	githubAppName: text("githubAppName"),
 	serverIp: text("serverIp"),
 	certificateType: certificateType("certificateType").notNull().default("none"),
 	host: text("host"),
-	githubClientId: text("githubClientId"),
-	githubClientSecret: text("githubClientSecret"),
-	githubInstallationId: text("githubInstallationId"),
-	githubPrivateKey: text("githubPrivateKey"),
-	githubWebhookSecret: text("githubWebhookSecret"),
 	letsEncryptEmail: text("letsEncryptEmail"),
 	sshPrivateKey: text("sshPrivateKey"),
 	enableDockerCleanup: boolean("enableDockerCleanup").notNull().default(false),
@@ -46,12 +38,6 @@ export const adminsRelations = relations(admins, ({ one, many }) => ({
 
 const createSchema = createInsertSchema(admins, {
 	adminId: z.string(),
-	githubAppName: z.string().optional(),
-	githubClientId: z.string().optional(),
-	githubClientSecret: z.string().optional(),
-	githubInstallationId: z.string().optional(),
-	githubPrivateKey: z.string().optional(),
-	githubAppId: z.number().optional(),
 	enableDockerCleanup: z.boolean().optional(),
 	sshPrivateKey: z.string().optional(),
 	certificateType: z.enum(["letsencrypt", "none"]).default("none"),
